Remove public_endpoint from keystone.conf template
The configuration parameter public_endpoint is required
only when base URL contains a prefix or endpoint is on
different host [1]. In this case even though the public
endpoint may contain prefix, the http request will land
on traefik or ingress where the prefix is handled and
the keystone application receives input requests without
any prefix. So there is no need to set this option.

Another reason for removing this option is, any incoming
requests on '/' are routed to public endpoint. So setting
health checks to http://localhost:5000 are routed to public
endpoint which is not expected.

This patch is to remove the public_endpoint option.

[1] https://docs.openstack.org/keystone/latest/configuration/config-options.html#DEFAULT.public_endpoint

Change-Id: I8e72f0fbecf9a80e70e5b2c09bbbdeb04265cd55