Switch glance to be in standalone mode.
Switch the glance service to be in standalone mode rather than
being server via mod_wsgi. This change is inline with upstream
documentation. "mod_wsgi: This deployment method is not recommended
for using Glance. The mod_wsgi protocol does not support
Transfer-Encoding: chunked and therefore makes it unsuitable for
use with Glance." *1. The documentation also explicitly recommends
standalone "...the Glance project team recommends that Glance be run
in its normal standalone configuration, particularly in production
environments."

When using mod_wsgi apache translates /<model-name> to /. In the
neutron charm this translation is handled by api-paste.ini. A
similar approach does not seem to work with glance, glance always
interprets the mode-name as a requested version number. To get
around this this change adds apache infront of glance to handle
the translation.

*1 https://docs.openstack.org/glance/latest/admin/apache-httpd.html#mod-wsgi

Change-Id: I013f1029c3b3f480b23c6bdf92eb155533c526b9