deps: add netifaces
Somewhere in the depths of the ceph-client ops interface code a
call occurs in charmhelpers that depends on netifaces; although
this is not actually needed for this k8s charm, it results in an
apt install of the python3-netifaces package.

Add this dependency to the charm requirements to avoid this
situation.

Change-Id: Ib78b187374e2c1fa01b5aada2daa538c975d27de