Tighten regex when looking for ovn db file
The payload containers sb, nb and northd have a shared db dir. The join
wrapper skips the join command if it detects an existing ovn db. But the regex
used by the join command is not specific enough. So the sb would skip the join
if the nb db was present. This patch tightens the regex so that the join
wrappers look for their own db.

Change-Id: I20b139244664a5218c236ec28756fefd617cc668