Use IP's instead of hostnames
When the OVN relay process starts early and the hostnames for the OVN
central units are not fully resolvable, the relay process gets stuck
in a state when it will never actually connect to the underlying OVN
central units.

Use IP addresses instead of hostnames to avoid this situation; this
does mean the relay will restart if the IP addresses change which
can and will happen during normal operation.

Closes-Bug: 2026596
Change-Id: Ib2cfa13f73182a3ab9cc37fc056fb78c972269a6