Update relation data from configure_charm
This change adds a mechanism for charms to be able to update
relation data as part of configure_charm rather than responding
to an explicit relation event.

The reason for this is that pod can be recycled at any point
causing it to loose its defered events. So, if a relation
adapter method to send relation data contains a defer that
deferred event can be lost and when a pod comes back it is
never resent.

Closes-Bug: 2036188
Change-Id: I2e6b27cdc05e93ebb25a7859b3c039ae64d82d9d