Share auth encryption key over relation
Currently the Auth encryption key for heat-api
and heat-api-cfn deployed via charm-heat-k8s
charm instances are different and so causes
issues when heat-engine from heat-api-cfn handles
request for heat-api. The heat-engines of both
heat-api and heat-api-cfn are used for handling
both heat-api and heat-api-cfn stacks as the AMQP
topic is same and not configurable.

Add a new interface heat-shared-config to share
the auth encryption key between heat-api and
heat-api-cfn.
heat-api update the relation data with auth
encryption key once the key is generated or
any new relations are connected via heat-shared-config.
Save the auth encryption key as juju secret
instead of plain text and grant access to
peer and units connected via heat-shared-interface.

Closes-Bug: #2036890
Change-Id: I2ec7f03b5c64d87585141e15b20b01172b14ecc4