Fix error when publishing charmcraft.yaml only charm
Since charmcraft 2.5, the files: metadata.yaml, actions.yaml, and
config.yaml have been merged into charmcraft.yaml.
The publishing job fails because it expects the metadata.yaml to be
present.
This change checks for which file to use.
The keys looked up in metadata.yaml for this specific task have the same
schema in charmcraft.yaml

Change-Id: I94ea3b8230e5dbf7d2474410fba7acc5e0836a63