Prevent unknown params passed to tempest charm
See: https://bugs.launchpad.net/juju/+bug/2051187

By default, juju allows extra params to be passed to an action.
This is risky here, because if a user makes a typo on a parameter,
it could be the difference between running a small fast set of tests,
or a very large slow set of tests...
For example, if the test-list or regex params are typoed.

Change-Id: Ic6daaab66b6187e8fb6151d861a0ba82e72ed756