Handle tempest resources clean-up
Three levels of clean-up is implemented to facilitate different
scenarios:
1. quick clean-up that removes resources created by tempest tests.
This can be used in-between test runs.
2. extensive clean-up that, in addition to 1, also removes test
accounts, project, and the associated network resources. This is
used after charm refresh where the previous test accounts becomes
obsoletes.
3. full cleanup that, in addition to 1 and 2, deletes the testing
domain, user, and project created through keystone relation. This
is used when relation to keystone breaks as well as when charm is
stopped or removed.

Change-Id: If172e9dde0e04c53849d7a0b30fc27f62c5bdbd9