Add functional test bundle for tempest-k8s
This is practically the same bundle and tests as the core test bundle,
but with the addition of the tempest-k8s charm.
This is to verify that the tempest-k8s charm deploys successfully,
and doesn't adversely impact the rest of the sunbeam deployment.

Also add the --create option to discover-tempest-config
so it can create flavors on the cloud for tempest to use.
(No flavors are available at deploy-time.)

Change-Id: If95cd4970960eacfe98ba93ad523d2c0cf6ccd25