Define concurrency for tempest run and accounts
Concurrency for tempest should be bounded to a small number (like 4),
to avoid overloading the cloud or runner machine,
but also this value should be bounded by the number of cpu cores.

For generating test accounts, this must always be >= the concurrency
selected for running tests.
Here we hardcode it to 16:
- It's risky generating a small number of accounts,
  because some tests require more than one.
- Having more accounts than needed helps mitigate the effects
  of credential locks not released
  ( https://bugs.launchpad.net/tempest/+bug/2052752 ).
- It's not worth scaling the number of accounts with
  tempest concurrency, because we want a minimum number of accounts
  generated regardless.

Change-Id: I96dbc65ca76a500651646cdd030912c73e1d86ce