Ensure that tempest init is retried
Sometimes the init process can fail, even though it checks if keystone
is ready first, because other cloud components may not be ready yet.
The solution for now is to retry the init process until it succeeds.
Retrying in this way can also be helpful in situations where
some configuration is causing the init process to fail,
allowing for debugging and reconfiguring until it succeeds.

Move logic from post_config_setup to configure_unit,
because post_config_setup doesn't have access to the event to defer.

Change-Id: Ie75d42a7344f466a6ef5f270bb22e0a8cf70efc7