Bump juju channel to 3.4
Move testing to juju 3.4

Set base to ubuntu@22.04 in all test bundles with traefik
charm to ubuntu@20.04. tox can be modified to include
default-series=;default-base=ubuntu@22.04 in TEST_MODEL_SETTINGS
but that is not working for lxd based cloud.

Update tox to use zaza from main instead of libjuju-3.1
for func tests.

Change-Id: I6c025e1350aad7ead922cf5feef19c67c308dbbd