Increase name complexity for tempest domain, user, and project
Since we are running clean-ups in the domain created by identity-ops
relation in multiple-stages of the charm's life-cycle, it would be
safer to use a more complex name for this domain. Otherwise, if
a user creates a domain also named "tempest" to do their own tests,
the charm will falsely clean-up resources there. The name changes
for admin user and project are simply for better readability.

Change-Id: Ief799048f1f261d052a02f3129b03626ba20c443