Remove cleanup on keystone relation broken
Cleanup at this stage in the lifecycle cannot work,
because it relies on credentials from keystone,
but keystone is gone away.

Move updating the tempest readiness from relation broken
to relation joined:
the charm doesn't need to consider re-initing tempest on relation gone,
but it does need to re-init on relation joined.

Also tidy and remove some unnecessary or dead code.

Change-Id: Id3cd9fb26cb7c0d661f11393d58508e50d481df1