Reduce the number of test accounts generated
With 16 as the value, and 3 networks on the sunbeam deployment,
about 150 ports are created, which is a lot and could cause issues
with quotas and other constraints.
Also, it can take a long time to create all the test account resources.

Reducing this number as much as possible will help address the above.

So here, halve the value to 8, which should still provide enough
test accounts with a reasonable buffer for the issue
with credential lock files not being cleaned up.
Tested once with the smoke tests (regex=smoke) and the refstack-2022.11
test list, and verified they can pass with this many test accounts.
(Note that the number of leaked lock files
is not always consistent though.)

Change-Id: I6087af68e162a5fc7351f023118194ca8a2d43ca