Upload charm before OCI images in publish charm job
Currently the publish-charm task uploads the OCI
image resources and then the charm. If a charm
introduces a new OCI image resource then this
workflow wont work as charmhub is not aware of
the new resource. Uploading charm first will
enable charmhub to understand the new resource.

Change tasks to upload charm first and then the OCI
image resources.

Change-Id: I70de4f19558c62caeaa8821a3bbd9cac1fd96c0d