Fix grid positions for hypervisor dashboard
There was empty space in the top left corner,
due to removal of panels during the porting process from the old dashboards.
This expands the top row to fill the full width, removing the empty space.

Also update the y coordinates of all the other panels,
since a horizontal panel was also removed in the porting process.
This didn't cause empty space, but better to update to be accurate.

Change-Id: I5bfaaa9d3ac5a2439ec8d92c83a9f2d4a1d2c3c4