[all] only publish active status at the end of the hook
The status_compound by designed set the unit's status many times during
hook execution. The unit's status is directly published to the
controller. This leads to outside observers seeing active status (and
lot of chatter around statuses) when the unit is in fact not ready.

With this change, units will only publish an active status at the end of
the hook execution. All other levels are still directly published to the
controller.
Units will no longer publish the WaitingStatus("no status yet"). This
creates a lot of chatter and holds little value.

Re-organize keystone __init__ not to publish false `Service not
bootstrap` status.

Closes-Bug: #2067016
Change-Id: Ie73b95972a44833ba4509f8fd2c2f52ed476004d