[keystone-k8s] configure charm when rotate id-svc fails
secret-rotate can be the first hook executed on certain failure modes,
such as the re-creation of a pod during a secret-rotate execution. This
can lead the keystone to be unconfigured + unstarted. Rotating
identity-service credentials necessitates keystone to be up and running.

Try to configure the charm if can't connect to keystone while trying to
create the new service account.

If after the new configure, the charm status is not active (happens when
a mandatory relation is missing for example), fail the event to be
retried later.

Closes-Bug: #2051257
Change-Id: I782232ca61c58f1040d1834c54f4c07c18a2e5d0