[all] reduce mysql-pooling to sensible defaults
Reduce the max_pool_size parameter from default 5 to 2, to reduce the
max number of connections each process needs.
Reduce connection_recycle_time from 200 to 20, to reduce life duration
of a connection, allowing faster recovery in case of failure of a
mysql-router.
Configure horizon-k8s to persist connections for 20 seconds, and enable
connection healthchecks to make sure the connection is usable.

Change-Id: Id88d5a8643b8dd2db618e18c8ad556f3e92282ba