Revert "Upgrade external libraries"
This reverts commit 6eb7f3b72b08d146df180063c8df4ac81f73a61d.

Reason for revert: Kubernetes service patch broke ovn-* relations

Upgrade to kubernetes_service_patch changed the way the service loadbalancer is setup. Before, it updated the application service, now it creates a new one. This broke ovn-* relations to outside the cluster because the `ingress-bound-address` is now filled with a k8s svc ip address, not reachable from the outside.

Change-Id: I38819cfcf647215e1eec595c2457300cb9058f90