Set tempest-ready to `False` when the unit start
The peer relation data `tempest-ready` persists across a reboot of the
host machine, and this can cause the inconsistency between the actual
pod state and the relation data. For example, relation data
`tempest-ready` can still be `True` before and after the host is
rebooted, but the tempest workspace directory created during tempest
init will be gone because the pod is recreated. So when the host is
rebooted we need to consider tempest to be no longer ready, so that in
the follow up config-changed hook, the charm will re-init tempest.

Bug: #2069767
Change-Id: I1db615e5094922b35da3b07552b0c045997abbfc