Removed `aggregates` label matcher and fix typo in sum by.
The `aggregate` in the sum by is a typo, and it should be `aggregates`. Also, with the `aggregates=~".+"` will show "No data" if there's no such label (this can happen when there is no aggregates in nova). Removing `aggregates=~".+"` will show the groups even if there is no aggregate (in this case it will be all running VMs)

Change-Id: I35f0ce74f79c81b2bb888483f7f94f719072c171