Fix Throughput Write query in compute dashboard
Now it matches the query pattern for Throughput Read.
Not sure why `juju_unit` was missing from the `on` grouping here;
it was present in the Throughput Read query.

This should fix a bug observed where
"found duplicate serious for the match group {domain="redacted"}
on the right hand-side of the operation".
Here, the `juju_unit` label was different.

Change-Id: I5cfe7704ac8d93c4f3ff50797ec2bae3ce6dc64b