[hypervisor] Add hostname as subject alternative names
Qemu blockdev-add command is unhappy when the hostname is not part of
the sAN in the certificates (even if it's already in the CN). Add
hostname to certificate's subject alternative names.

If there is a migration address, it will be the prefered one for qemu
migration, therefore also include hostname exposed on this address.

Change-Id: I7a1f0e9e0a21f8dbc4bab94acec4f1c5b445a054
Signed-off-by: Guillaume Boutry <guillaume.boutry@canonical.com>