keystone: always log CADF audit messages and simplify workfow
Right now, the CADF audit messages will not be logged unless
telemetry is enabled.

oslo.messaging allows multiple drivers to be used, so we'll
just use the "log" driver to always log CADF events and
enable the "messagingv2" driver if telemetry is enabled.

This means that we can drop our custom amqp listener,
simplifying the charm.

The tests are updated accordingly, testing audit log messages
with and without the telemetry feature.

Change-Id: I0bcdcd8b5477d04c6bd0d8d5cc691e4b5ca7edcf