Update keystone logging fmt
We'll update the Keystone logging format to match that
of other services, simplifying log parsing (e.g. extracting
audit information).

We're also adding a functional test to ensure that the log format
is consistent.

While at it, we're adjusting the test order as a workaround for
the following bug: https://bugs.launchpad.net/sunbeam-charms/+bug/2114836

Change-Id: Ib585e0925c89771d7221f547eda04b19633a8ce8