Mitigate flaky Keystone audit test
The Keystone audit test flips the "enable-telemetry-notifications"
setting to ensure that the audit messages are emitted with
and without telemetry notifications enabled.

As a result, Keystone gets restarted twice and it can take
a while for the service to recover, see the following bug:
https://bugs.launchpad.net/sunbeam-charms/+bug/2114836

We're getting Keystone 503 errors, despite the fact that the
unit is reported as ready.

To mitigate this and avoid test flakes, we'll simplify the
test and avoid changing this charm setting.

Change-Id: I3a0d72a2383188ce3c7408e5c1dcd5956f7baedd
Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>
(cherry picked from commit 4449f1189982d95b81ade324a9d6dd1e55941e34)