Fix config typo (api_paste_confg)
The config templates have a typo (api_paste_confg) that went
unnoticed since we were using the default paste file location.

We'll go ahead and drop it since some of the projects have renamed
it and we're using the default path anyway. The middleware tests
will catch it if the paste pipeline is no longer applied.

Change-Id: Ia736740ee6f2ccb0c86d4110c17d621f107a9a27
Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>