[openstack-exporter-k8s] Correct dashboards when having Ceph-shared storage backend
Add logic to some Openstack-Exporter-Operator dashboards to correctly
report storage status when Ceph is used as the storage backend and
no ephemeral storage available. The issue is reported in GitHub:
https://github.com/canonical/openstack-exporter-operator/issues/154

This patch add a workaround logic to detect whether the Openstack
storage is likely backed by Ceph and nodes' ephemeral storages aren't
used. Calculations are also updated for several dashboard parameters
such as total storage or free storage, so that they are correctly
properly when the condition is met. If the condition is not met, the
logic will simply fallback to existing expression.

A more detailed explanation and testing proof can be found in:
https://github.com/canonical/openstack-exporter-operator/pull/157

Change-Id: Id14fd0acd3f5b9601a08be35cc078f00d8035ef8
Signed-off-by: Quang Ngo <quang.ngo@canonical.com>