openstack-hypervisor: allow passing dpdk ports
The openstack-hypervisor snap can now receive a list of dpdk ports
that will be:

* bound to the dpdk compatible driver (defaults to vfio-pci)
* added to bonds and bridges based on the existing maas/netplan
  configuration

https://github.com/canonical/snap-openstack-hypervisor/pull/111

The charm will receive the list of dpdk ports through the
"set-hypervisor-local-settings" action using the "ovs-dpdk-ports"
key.

The OVS pinned cpus and huge pages will then be reconfigured based
on the NUMA nodes of the specified interfaces.

While at it, we'll ensure that the system ovs-dpctl binary is used
instead of the snap alias when clearing system data paths.

Change-Id: I727b49d1df9d77a9d4ee51ac11aa264806f662b3
Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>