ironic-conductor-k8s: Adds set-temp-url-secret action
The action is meant to be run on the leader after the mandatory
relations are ready. The action will check that the Glance and
Swift endpoints exist, and that Glance has the Swift backend
configured. The action will store the "x-account-meta-temp-url-key"
in Swift, if needed.

Changes from the original commit:
- Updated pyproject.toml constraint-dependencies to the 2024.1
  upper-constraints.
- Updated charms/ironic-conductor-k8s/uv.lock (ran `uv lock`).

Change-Id: I4021ba1be507251219ae821eb8f840ca5ed97f32
Signed-off-by: Claudiu Belu <cbelu@cloudbasesolutions.com>
(cherry picked from commit 811e30dfcab21bea7107351b2def2e85382fd296)