Allow hypervisor and cinder-volume to be installed in devmode
Temporarily, while waiting for SnapD interface to be available, allow
turning on/off devmode on the snaps for both cinder-volume and
openstack-hypervisor.

Unfortunately, the logic is a bit complex, since, if the `refresh`
commands does not ensure an actual change of revision the `devmode`
parameter will be ignored. In this case, the snap must be removed and
re-installed. While this logic is complex, it will only be triggered
in confinement changes + ~latest-ness.

OpenStack-hypervisor is reimplementing the logic because it does not
derive yet from the charm snap base class.

Change-Id: If2995fae1924f42b1efc2d583aeaffba7d0c6b73
Signed-off-by: Guillaume Boutry <guillaume.boutry@canonical.com>