neutron-k8s: Fixes ironic-neutron-agent service state
Currently, the NeutronServerPebbleHandler starts all the container
services, including the ironic-neutron-agent service, even if the
ironic-api relation is not present. This causes the neutron-k8s unit
to remain in a "waiting" state, as not all its container services
enter an active state.

This amends the pebble handler's behaviour, starting the
ironic-neutron-agent service only if the ironic-api relation is
present, and stopping it otherwise. Additionally, service_ready
will now discount the ironic-neutron-agent if the ironic-api
relation is not present.

Change-Id: I7dfe8a59392fbbd20bed103309bb954fda65d491
Signed-off-by: Claudiu Belu <cbelu@cloudbasesolutions.com>