Upgrade traefik-route after upstream move
If the traefik-route-internal relation is the last relation added, the
nova-k8s and heat-k8s charms are blocked in a Waiting state.
Updates the traefik-route-internal handler for these charms to
self.configure_charm. The new handler calls self.handle_traefik_ready,
and it will allow the charm to enter an active state.

Updates their unit tests to check against this issue.

Change-Id: I0b751cfb1bcd99e1e87c40c60dea69717ba9fae9
Signed-off-by: Guillaume Boutry <guillaume.boutry@canonical.com>
Signed-off-by: Claudiu Belu <cbelu@cloudbasesolutions.com>