ci: Increase timeouts
Currently, the func-test-openstack fails due to timeouts. In the logs,
we can see that zaza (zaza.charm_lifecycle.deploy) waits for the deployment
to settle for 3600 seconds (default TEST_DEPLOY_TIMEOUT value):

[INFO] Timeout for deployment to settle set to: 3600

We are deploying more charms, so they take longer to settle.

The sos networking plugin tends to hit the 300 second timeout.
Increasing it to 600.

Change-Id: I18f1a0dd7283a26de7d28a25eafbb96515dcc4ed
Signed-off-by: Claudiu Belu <cbelu@cloudbasesolutions.com>