Allow using Keystone from a separate region
In multi-region environments, Keystone services will reside
in a separate region.

As such, we need to specify the Keystone region in the
[keystone_authtoken] section, otherwise the API services
(such as Nova) will expect to find the identity service in the
same (secondary) region, failing to authenticate user tokens:

  nova.api.openstack keystoneauth1.exceptions.catalog.EndpointNotFound:
  internal endpoint for identity service in SecondRegion region not found

We'll fetch the Keystone region from the identity relation data.

Change-Id: I818e1d5964819cdf3c92585fbd9544a8d9ce10b6
Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>