horizon: multi-region support
The list of regions must be specified in the Horizon configuration,
otherwise the user won't be able to switch between regions.

We'll update the Horizon charm to fetch the list of regions from
Keystone using the identity-ops interface, store it in the app peer
data and then set the "AVAILABLE_REGIONS" Horizon option.

We're also adding an action to trigger a refresh, which is expected
to be called during the secondary region bootstrap.

Change-Id: I1dc95aba7874981de1f9573815830cde2273e048
Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>