neutron:k8s Adds functionality to set enable-distributed-floating-ip
This adds a code to enable enable-distributed-floating-ip.
This parameter was earlier hardcoded to false. With this
change user will be able to set value manually.

Note: Default value is still set to False as placeholder.
This needs to be decided if default value has to be changed
to True or keep as False.

enable-distributed-floating-ip:
   default: false
   description: |
      Enable or disable distributed floating IP functionality in OVN.
      When enabled, floating IPs are managed by distributed routers.
   type: boolean

Closes-Bug: 2129974
Change-Id: Ic06d6cd33a87e709aadd9e157c94c23b927b7a71
Signed-off-by: sikander143 <sikander.singh@canonical.com>