fix TlsCertificateHandlers to pass private key as secret
In get_certificate_context, pass the secret if of the
private key instead of key itself in TlsCertificateHandler
relation handler.

In addition, modify the logic for relation handler ready
for TlsCertificateHandler. get_certs() for non-leader unit
wont return any certificates and so just return True without
checking for certificates None or not. It is ok to do this
as leader unit when it is ready will have the certificates
updated in relation app data.

Change-Id: I40d064a8288d3a03eb8e9cb6c73c436721c13059
Signed-off-by: Hemanth Nakkina <hemanth.nakkina@canonical.com>