glance-k8s: Fixes swift configuration section
According to [1], the current swift configuration is deprecated, and having
a Swift back-end configuration file is the recommended way. The Swift
back-end configuration file also contains the project_domain_name and
user_domain_name config options, which are needed for the keystone
client used when querying for the swift keystone endpoint.

[1] https://docs.openstack.org/glance/2025.1/configuration/configuring.html

Change-Id: I08ec5d6df86575769443c60e6ce3203fafef0c55
Signed-off-by: Claudiu Belu <cbelu@cloudbasesolutions.com>