ironic-conductor-k8s: Fixes ironic-conductor charm functionality
The ironic-conductor's HTTP and TFTP services are meant to be accessible
by the bare metal nodes. As such, they need an IP through which they can
be reached. For this, we can assign an External IP (by using a Loadbalancer
service).

Remove the ipxe-http-port config option, use port 80 instead.

Start the HTTP and TFP services.

Set the "image_download_source = swift" and ipxe_use_swift = true config options,
in order to use the swift image source, instead of the ironic-conductor service
pulling the images and serving them through the HTTP service.

Change-Id: Iaa14a16b74d24468a406ceb33f8eb69e9a8acdd1
Signed-off-by: Claudiu Belu <cbelu@cloudbasesolutions.com>