Fix getting sans ip addresses
Currently the sans ip addresses for an unit is
computed based on bind and ingress IPv4 addresses
for all the relations that are integrated.
However if a new relation is integrated, the
sans ip addresses are recomputed and can have
extra IP addresses.
If the certificates are generated with old sans ip
addresses, then it is not possible to retreive the
certificate from TLS certificate interface v4 library
when the sans ip addresses are changed. sans ip
addresses used for certificate generation should
be fixed and so all the necessary ip addresses
are to be retrieved before in hand.

To reoslve this, Compute sans ip addresses based on
all the relations defined rather than that are integrated.

Change-Id: I5513de1a677fa8bfb0bbb2cd65d7f9d6c7d846f2
Signed-off-by: Hemanth Nakkina <hemanth.nakkina@canonical.com>