Add openstack-hypervisor migration ip to sans ip
The migration binding address is added to sans
domain name but not to the sans ip.
Add migration binding address to sans ip and
provide the same for certificate generation.

To regenerate certificates on existing deployments,
a new function is added to reverify the generated
sans ip, sans dns with the one in relation data
as part of CSR. If they are mismatched, update
the request certificates with updated sans ip,
sans dns and regenerate the certificates.
Invoke the above function on upgrade-hook charm.

The regeneration logic is also required for
neutron-k8s, octavia-k8s, ovn-central-k8s,
ovn-relay-k8s and sunbeam-clusterd so that
the certificates are regenerated on charm
upgrade (to handle older deployements which
uses tls certifiates interface v3 to newer
tls certificates interface v4)

Closes-bug: https://bugs.launchpad.net/sunbeam-charms/+bug/2133177

Change-Id: I04b95a6d00d9e0ace6f277d05e42853297b326e2
Signed-off-by: Hemanth Nakkina <hemanth.nakkina@canonical.com>